requirements.txt
pystray==0.19.5
PyYAML==6.0.1
pillow==10.3.0
keyboard==0.13.5
mss==9.0.1
numpy==1.26.4